move generic Api superclass from url.ast.data.path to url.ast.api
This is used in url.ast.catalog, data, and model and so is not
specific to data paths.
